requirements.txt
matplotlib==3.9.3
numpy==2.2.0
pandas==2.2.3
Pillow==11.0.0
scikit_learn==1.6.0
torch==2.5.1
tqdm==4.67.0
